moved Errors to ArpaEtlErrors
